[WRAPPER] Fix some logic errors in #4214 (#4264)
1. Add Xt returned error-handler conversion.

2. Fix FUSE stat and mode argument conversions.

3. Fix llrintl return values.

4. Add the `user_data` argument to PCRE2 callouts.

5. Add the NSS UCS2-to-UTF8 callback wrapper.

6. Add the whence argument to TIFF seek callbacks.

+++ src/wrapped/wrappedlibm.c
@@ -234,15 +234,15 @@ FROUND(llrintf, float, long)
 FROUND(nearbyint, double, double)
 FROUND(nearbyintf, float, float)
 #ifdef HAVE_LD80BITS
-FROUND(llrintl, long double, long double)
+FROUND(llrintl, long double, long long)
 #else
-EXPORT double my_llrintl(x64emu_t* emu, double val)
+EXPORT int64_t my_llrintl(x64emu_t* emu, double val)
 {
     if (BOX64ENV(sync_rounding)) {
         int round = emu->cw.x16 & 0xc00;
         fesetround(TO_NATIVE(round));
     }
-    return llrint(val);
+    return (int64_t)llrint(val);
 }
 EXPORT double my_nexttoward(x64emu_t* emu, double val, double to)
 {
